boxes/tests/factories.py
import factory
from django.contrib.auth.models import User


class UserFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = User
        django_get_or_create = ("username",)

    is_staff = True
    password = "yxcv1234"
    username = "test-user"
