requirements/base.txt
asgiref==3.3.1
Django==3.1.5
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
django-extensions==3.1.0
Werkzeug==1.0.1 # required for django-extensions
django-guardian==2.3.0
django-crispy-forms==1.10.0